+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+#### v1.0.8
+
+- Implemented `ngzFilter` pipe for generic filtering
+- Implemented `ngzSort` pipe for generic ordering
+- Implemented `ngzAsDate` pipe for converting date partials into a Date instance
+
 #### v1.0.7
 
 - `ngz-grid > ng-container[ngzGridColumnDef][hasOrdering]` now supports values of `'asc'` and `'desc'` to provide default sorting by this column